compare gene trends between branches
This commit enables tthe comparison while using the new implementation
that uses mellon for gene trends and saves all results in the procided
Anntata.

To finilize the new workflow of passing on AnnData I udpated all
functuions used in the tutorial and updated the tutorial itself
with the new workflow.

Since we decided to remove unused functions such as run_tsne I also
removed the following functions that are covered by scanpy:
 - palantir.utils.determine_cell_clusters
 - palantir.plot.plot_cell_clusters